Final step in current database cleanup process
This push went through the 3.5 rules database utils. Next up will be
more feature progression, possibly going to the next chapter in the book

+++ project/app/src/main/java/com/brave_bunny/dndhelper/database/inprogress/InProgressUtils/InProgressDomainsUtil.java
@@ -54,7 +54,7 @@ public class InProgressDomainsUtil {
         deleteFromTable(context, query, selectionArgs);
     }
 
-    public static void removeDomainSelection(Context context, long rowIndex, int domainId) {
+    public static void removeDomainSelection(Context context, long rowIndex, long domainId) {
         String query = characterIdLabel() + " = ? AND " +
                 domainIdLabel() + " = ?";
         String[] selectionArgs = new String[]{Long.toString(rowIndex), Long.toString(domainId)};
@@ -72,7 +72,7 @@ public class InProgressDomainsUtil {
         }
     }
 
-    public static void addDomainSelection(Context context, long rowIndex, int domainId) {
+    public static void addDomainSelection(Context context, long rowIndex, long domainId) {
         ContentValues values = new ContentValues();
         values.put(characterIdLabel(), rowIndex);
         values.put(domainIdLabel(), domainId);
